siljeon2/SilJeon/your_voice/backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
tensorflow==2.16.1
Flask-Cors==4.0.1
pydub==0.25.1
numpy==1.26.4
librosa==0.10.2.post1
soundfile==0.12.1
pymysql==1.1.1
dbutils==3.1.0
google-auth==2.29.0
gunicorn